standardize todo formatting for backend
there were some inconsistencies with how a todo dataObj needs to be.
specifically with date.

now there doesn't need to be a 'dateSpecifiedByUser' property.
This will be made depending on how to todo is delivered.

the todo needs either: a "urgency" property (value 1, 2 or 3)
or a "lastDayOfDeadline" property (value is date)