expand UIStates classes
the classes are still unfinished, but now the available buttons are added.
What is difficult is to fill in the todoDisplay container.
It seems I have managed by now, so that is a small victory.

What did I do:
todoDisplayContent is  function which, if ran, will fill todoDisplay apprioprately
I saved domElements.todoDisplayContent as  variable. no "()".
I bind domElements to this variable, otherwise the "this" keyword doen not work in the function

then I created a function "updateTodoDisplayContent" on the instance,
there I call this variable as a function.

now the todoDisplay can be update by running "updateTodoDisplayContent" on the instance