remove overview class
It was not well defined yet. I removed it.
in a few places there is now the following call: frontendController.applyEisenhowerMatrix()
is I ever add a dynamic overview state class I can replace those with something like:
overviewStates.currentOverview.apply()

so I do not yet see big trouble for expanding this code in that way in the future.
right now it is very difficult for me to see how I would implement this as a class in the future.
So will leave it simple and expandable for now.